pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "svg-pipeline"
version = "0.1.0"
description = "Transform SVG/image sources into all required derivative formats"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "SVG Pipeline Contributors" }
]
keywords = ["svg", "favicon", "icons", "image-processing", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

dependencies = [
    "pillow>=10.0.0",
    "cairosvg>=2.7.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
opencv = ["opencv-python>=4.8.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
svg-pipeline = "svg_pipeline.cli:app"

[project.urls]
Homepage = "https://github.com/svg-pipeline/svg-pipeline"
Documentation = "https://github.com/svg-pipeline/svg-pipeline#readme"
Repository = "https://github.com/svg-pipeline/svg-pipeline"

[tool.hatch.build.targets.wheel]
packages = ["src/svg_pipeline"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
